Split gpu/analyzer.rs (2,625 lines) into per-evaluation modules (#520)
Extract five evaluation-specific modules from the monolithic analyzer.rs:
- helpful_evaluation.rs: helpful synapse pipeline + batch evaluation
- harmful_evaluation.rs: harmful synapse pipeline + batch evaluation
- relu_evaluation.rs: ReLU activation pipeline + evaluation
- activation_evaluation.rs: activation function pipeline + single/batched evaluation
- bias_evaluation.rs: bias pipeline + grid search evaluation

analyzer.rs now contains only core shared logic (492 lines): GpuAnalyzer struct,
GpuEvaluator trait, initialisation, system checks, and helper functions.
